# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Healthcare Recruiting ATS/CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  role: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface PasswordInput {
  currentPassword: string;
  newPassword: string;
}

export interface DashboardStats {
  totalCandidates: number;
  newToday: number;
  duplicatesFound: number;
  resumesUploaded: number;
  activePipeline: number;
  totalThisWeek?: number;
  totalThisMonth?: number;
}

export interface BreakdownItem {
  label: string;
  count: number;
}

export interface WeeklyGrowthItem {
  week: string;
  count: number;
  date?: string;
}

export interface Tag {
  id: number;
  name: string;
  /** @nullable */
  color?: string | null;
}

export interface Candidate {
  id: number;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zipCode?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  experience?: string | null;
  /** @nullable */
  licenseType?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  preferredLocation?: string | null;
  /** @nullable */
  availability?: string | null;
  /** @nullable */
  currentStatus?: string | null;
  /** @nullable */
  pipelineStage?: string | null;
  /** @nullable */
  recruiterNotes?: string | null;
  /** @nullable */
  resumeFile?: string | null;
  /** @nullable */
  source?: string | null;
  isDuplicate?: boolean;
  tags?: Tag[];
  createdAt: string;
  updatedAt: string;
}

export interface CandidateInput {
  fullName: string;
  phone?: string;
  email?: string;
  city?: string;
  state?: string;
  zipCode?: string;
  position?: string;
  specialty?: string;
  experience?: string;
  licenseType?: string;
  licenseNumber?: string;
  preferredLocation?: string;
  availability?: string;
  currentStatus?: string;
  pipelineStage?: string;
  recruiterNotes?: string;
  source?: string;
}

export interface CandidateUpdate {
  fullName?: string;
  phone?: string;
  email?: string;
  city?: string;
  state?: string;
  zipCode?: string;
  position?: string;
  specialty?: string;
  experience?: string;
  licenseType?: string;
  licenseNumber?: string;
  preferredLocation?: string;
  availability?: string;
  currentStatus?: string;
  pipelineStage?: string;
  recruiterNotes?: string;
  source?: string;
}

export interface CandidateListResponse {
  data: Candidate[];
  total: number;
  page: number;
  limit: number;
  totalPages?: number;
}

export interface StageInput {
  stage: string;
}

export interface BulkIdsInput {
  ids: number[];
}

export type BulkUpdateInputUpdates = {
  pipelineStage?: string;
  currentStatus?: string;
  specialty?: string;
};

export interface BulkUpdateInput {
  ids: number[];
  updates: BulkUpdateInputUpdates;
}

export interface DuplicateGroup {
  duplicateKey: string;
  duplicateType?: string;
  candidates: Candidate[];
}

export type DuplicateActionInputAction = typeof DuplicateActionInputAction[keyof typeof DuplicateActionInputAction];


export const DuplicateActionInputAction = {
  ignore: 'ignore',
  update: 'update',
  merge: 'merge',
} as const;

export interface DuplicateActionInput {
  action: DuplicateActionInputAction;
  targetId: number;
}

export interface Note {
  id: number;
  candidateId: number;
  content: string;
  /** @nullable */
  noteType?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface NoteInput {
  content: string;
  noteType?: string;
}

export interface NoteUpdate {
  content?: string;
  noteType?: string;
}

export interface Activity {
  id: number;
  /** @nullable */
  candidateId?: number | null;
  type: string;
  description: string;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export interface ActivityInput {
  type: string;
  description: string;
  metadata?: string;
}

export interface TagInput {
  name: string;
  color?: string;
}

export interface PipelineColumn {
  stage: string;
  label?: string;
  count?: number;
  candidates: Candidate[];
}

export type ImportInputRowsItem = { [key: string]: unknown };

export interface ImportInput {
  rows: ImportInputRowsItem[];
  filename?: string;
}

export interface ImportResult {
  imported: number;
  skipped: number;
  duplicates: number;
  errors: string[];
  /** @nullable */
  importLogId?: number | null;
}

export interface ImportLog {
  id: number;
  filename: string;
  imported: number;
  skipped: number;
  duplicates: number;
  /** @nullable */
  errors?: string | null;
  createdAt: string;
}

export type ListCandidatesParams = {
page?: number;
limit?: number;
search?: string;
specialty?: string;
state?: string;
status?: string;
licenseType?: string;
pipelineStage?: string;
sortBy?: string;
sortOrder?: ListCandidatesSortOrder;
};

export type ListCandidatesSortOrder = typeof ListCandidatesSortOrder[keyof typeof ListCandidatesSortOrder];


export const ListCandidatesSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type SearchCandidatesParams = {
q: string;
page?: number;
limit?: number;
};