# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Healthcare Recruiting ATS/CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: number;
  /** @nullable */
  candidateId?: number | null;
  type: string;
  description: string;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}