# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Healthcare Recruiting ATS/CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  role: string;
}